Refactor publishing-related code to use pipelines
- Removed references to Aspire.Hosting.Publishing in multiple files and replaced them with Aspire.Hosting.Pipelines.
- Introduced new classes: CompletionState, IPipelineActivityReporter, IPublishingStep, IPublishingTask, and their implementations.
- Created a NullPublishingActivityReporter for testing purposes.
- Added PipelineActivityReporter to manage publishing activities.
- Implemented PipelineContext and PipelineStepContext for better context management during pipeline execution.
- Updated tests to reflect changes in namespaces and class references.
