+++ src/Aspire.Hosting/Pipelines/CompletionState.cs
@@ -3,7 +3,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Represents the completion state of a publishing activity (task, step, or top-level operation).
