+++ src/Aspire.Hosting/Pipelines/IPipelineActivityReporter.cs
@@ -3,7 +3,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Interface for reporting publishing activities.
