+++ src/Aspire.Hosting/Pipelines/NullPipelineActivityReporter.cs
@@ -5,7 +5,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// A no-op implementation of <see cref="IPipelineActivityReporter"/> for testing purposes.
