+++ src/Aspire.Hosting/Pipelines/PipelineStepContext.cs
@@ -2,10 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics.CodeAnalysis;
-using Aspire.Hosting.Publishing;
+using Aspire.Hosting.ApplicationModel;
 using Microsoft.Extensions.Logging;
 
-namespace Aspire.Hosting.ApplicationModel;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Provides contextual information for a specific pipeline step execution.
