+++ src/Aspire.Hosting/Pipelines/PublishingTask.cs
@@ -5,7 +5,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Represents a publishing task, which belongs to a step.
